Add support for cs2_gk.dat (global key) to catsys
* cs2_gk.dat is an older key file that doesn't seem to be used anymore.
* It has only been found in Djibril4 and Campanella trials so far (dated 2008 - 2010)
* The key uses a simple round of Blowfish encryption over an MD5 hash of the target file. The key for the Blowfish cipher is the V_CODE (1) seed (not even an ounce of Mersenne Twister... *gasp*)
* The global key stores any number of files with encrypted MD5 hashes, but most often its seen holding ["<cs2game>.exe", "install.exe"], in that order.

Commentary:
The actual discovery was made from ac.exe (animation compiler) of all things. Not the Cs2 executables. This may be due to ac.exe being compiled in DEBUG mode, making it much easier to read with Ghidra. The key is read by dividing the key file length by 0x30 (48), and then reading the first 0x20 (32) bytes for the filename, and next 0x10 (16) bytes for that files encrypted MD5 hash. That hash is then decrypted using Blowfish initialized with the V_CODE (1) seed.

This seed was actually guessed, as none of the normal Cs2 operations occur in ac.exe, the function that *would* call the key reading function, (and pass in the V_CODE seed) was just sitting there chillin'. Nobody calling it... poor thing :c